requirements.txt
pystray==0.19.4
keyboard==0.13.5
Pillow==9.1.1
customtkinter==5.0.3